Created Products service to call from the controller to get the products, which then calls Products data access to retrieve products.
Created exchange rate data access and currency service. Currencies are currently hardcoded in Currencies file, but these could potentially be stored in a database.

Exchange rates are retrieved from (fake) database using data access class, but these could potentially be retrieved from some API, in which case an ExchangeRateAPIService could be created (implementing ICurrencyService interface). The implementation of ConvertCurrency method is likely to be the same, so a base class could be created with a virtual method ConvertCurrency to provide default functionality that could be overriden if needed.

Set nullable setting to enable nullable reference types, to avoid having to manually add a lot of null checks and to automatically enforce types such as strings that should not be null (or make them nullable where null values are allowed).

Added logging and exceptions in few places and comments on top of the methods.

Implemented unit tests.